Use a len attribute instead of the __len__ method
On 32-bit architecture, the size of the integer returned by the __len__
method are restricted to the maximum size of an integer in C. Any number
that is larger than 2**32 results in an OverflowError.

requests uses a utility method to determine the length of a body of a
request. By setting `len` as an attribute, we can still take advantage of
super_len's behaviour and we can hack around the OverflowError.

Closes #80