Fix authentication handlers using digest auth
Requests 2.8.0 introduced a thread-safe HTTPDigestAuth class which was
backwards incompatible for anything that attempts to work with it. This
adds a compatibility layer for convenience based on the version of
requests that is installed alongside the toolbelt.

This also updates the HTTPProxyDigestAuth class to be threadsafe by
utilizing the thread local storage when it is available.

Closes #108