Add more updates
While looking at the tests, I realized that there was no way this was
an actual SNI adapter. This does SSL verification plain and simple.

Since it is *not* specific to SNI, I've renamed it once more and updated
references of SNI around the feature to SSL for consistency, clarity, and
accuracy.