Do not fail on unhandled ImportErrors
In gh-121 we realized that some people are ignoring documentation and
attempting to use adapters that only work with more recent versions of
Requests and urllib3. As such, we need to handle ImportErrors from
our compatibility module better and issue warnings closer to where a
user might be encountering the problem.

Closes gh-121