Work around MRO when monkeypatching; thx @Bogdanp!
The MRO of requests.adapters.HTTPAdapter is recomputed when we
monkeypatch it. At the time it is recomputed, it does not hold a
reference to the original HTTPAdapter (as is the case when the class
relationship is determined, at import time). Instead, the bases are
dynamically evaluated; at this time the desired base class has been
replaced by the monkeypatch.

This works around that by using an additional class that is injected
into the MRO of both the initial adapter and the one we are going to
monkeypatch in. We also force this class to only call super() once via a
flag (the flag is kept as an instance attribute copied from the class). In this
way the patch class becomes a base for both adapters, and the MRO ends up being
computed properly.