Set SNI to Host header in HostHeaderSSLAdapter
Modern TLS/SSL servers can serve content and certificates based on the
Server Name Indication (SNI) that is sent in the Client Hello. By
setting the server_hostname in the connection pool arguments, the
underlying libraries will send the provided hostname as SNI. This allows
us to use the HostHeaderSSLAdapter to request content from a specific IP
address, while still requesting the server to sent content based on a
hostname.